major rework: check ro in parent ns
When remounting read-write, add code to verify that all mount points are read-only in the parent mount namespace. This makes sure that rofairy does not inadvertantly exposes a filesystem as writable system wide.

Give up on `set -eu` because it is getting in the way too much; it breaks the script  when a loop finishes with a failing condition.